Created Current, Legacy and Retired
"Current" code should be the code that is run on the robot. One should be able to simply download the "Current" folder and move all the scripts to the appropriate locations to expect a functional robot.

"Legacy" contains old code or test scripts that for whatever reason still need to be used.

"Retired" contains code that is not used in the current project but may need to be referred to.